Make side castellations asymmetric
The long sides could previously fit in either way around.  This was
originally because it didn't matter which way around they went.
It now matters because the shield does not extend the full length of
the ruggeduino, so the notches that emerge from the tops of the sides
are actually difference widths.

This patch extends one of the 8mm notches on each side to 10mm so that
the sides only fit in one way around.