Add two standoffs to flaily bits
The flaily bits were too flaily.  So this patch adds a standoff at the
end of each flaily bit.  It keeps the double-layer lid situation,
however this may not really be necessary.  We'll see.

In order to fit the standoffs, the lid had to be extended lengthways
by another 0.5mm.  Since there is now space, the standoff that was
between the USB and power connectors now no longer has to intersect
the end plate.  So that end plate is now one piece again rather than
two.